Refactor Blackjack module structure by moving DTOs and enums to a dedicated models directory. Added new request and response DTOs for game actions.
