+++ src/modules/blackjack/models/interfaces/game-state.interface.ts
